Point `README.md` section `Status` to `wip.md`.
`README.md` is old. It has a lot of useful info but it needs editing to
make it useful for day-to-day work. So now it points to `wip.md` for
day-to-day task tracking.

I track tasks manually in `wip.md`. I kept a *WIP-and-RIP* filtered copy
of the Git log (the output of `;glw`) in `WIP-log.md`. Erase this
document when I finish all the tasks listed in it.

I was tracking tasks in the Git log, but as Sean pointed out this is not
practical since editing the log to mark old tasks as done requires a
rebase. I retired shortcut `;glw`.