Implement API call `SetBridgeLED`
`SetBridgeLED` behavior is pinned down by unit tests. The unit tests
reflect the behavior described by the `SetBridgeLED` entry in the
UsbProtocol table.

`SetBridgeLED` reads a two-byte payload, so it was harder to test than
`GetBridgeLed`. `GetBridgeLED` reads a single byte of payload. The
fake `UsbReadN` simply read from the fake hardware pin register. The
test writes to this fake hardware register before calling `GetBridgeLed`.

To fake reading multiple bytes, I created an alternative fake for
UsbReadN that reads from a global array. The test is responsible to load
the global array before calling `SetBridgeLED` (otherwise the values in
the array are unpredictable).