Fix Queue to work in ISR
Queue passes unit tests. Queue passes system tests for USBProtocol
command GetSensorLED.

GetSensorLED has a single-byte parameter. To send commands over SPI with
multi-byte parameters, I need to buffer more than one byte on the Sensor
board. I need a Queue: a FIFO Buffer for SPI bytes coming into the
Sensor.

Any call to the Queue within the interrupt routine resulted in some very
strange behavior. I thought maybe the Queue was corrupting memory
somehow.

No. It turns out my Queue lib was fine. The moment I added function
calls to my interrupt routine, I introduced a bug. I eventually realized
it didn't matter what kind of function call I made.

Adding even a single function call in my interrupt routine, there was a
huge overhead to store register values before the first line of C code
in the routine was executed. This created a problem when the Sensor
wanted to send data. The Sensor drives DataReady LOW to request the
Master execute a transmission, and resets to HIGH when the transmission
is done. The Sensor drives DataReady HIGH in the ISR, but it took so
long for the ISR overhead, the Master sampled a second DataReady LOW
before the ISR had a chance to drive DataReady HIGH. So the Master went
ahead and read a byte the Sensor was not ready to send. And now I know
the Sensor sends 0xFF if I don't write a value to the SPI data register.
There is no way to drive DataReady HIGH any faster unless I wanted to
add a timer interrupt, but that is overkill. I just needed more robust
signaling between SPI Master and Slave.

The solution was to add a synchronizing check for HIGH on the Master.
This throttles the Master when Slave sends multiple bytes to Master.
That handles most cases, but there is a corner case when the Master
follows receiving from the Sensor with writing to the Sensor. Now that
the Master checks for HIGH before checking for LOW, I was able to move
DataReady HIGH out of the ISR, into the same section of code that drives
DataReady LOW. So every byte sent by the Sensor is sandwiched between a
LOW/HIGH on DataReady, and the Master is prevented from starting any new
SPI communication until it sees DataReady go HIGH after the final byte
sent.

I simplified the memory management for the Queue lib and cleaned up the
unit tests to match the new memory setup.