vis-spi-out includes SensorVis.c to optimize assembler for I/O registers
SensorVis includes functions I want to inline. The bodies are in
SensorVis.h. This allows avr-gcc to inline the function when it links
vis-spi-out.o with SensorVis.o.

But avr-gcc cannot optimize inline functions that operate on the I/O
registers. Hardware definition headers are only included in
vis-spi-out.c. Similarly, the fake hardware is only included in the test
runner. With the hardware definitions unknown at compile time, avr-gcc
has to use the most generic data load and data store instructions,
instead of register I/O specific instructions sbi, cbi, sbis, sbic for
bit manipulation, and in/out for manipulating the entire value. This
adds a lot of overhead and noticeably slows execution, even to the point
that the microcontroller cannot read the ADC fast enough when getting
a frame of photodiode array readings.

Inability to optimize due to abstraction for better code quality is a
common problem in embedded code. Use inline rather than macros to
overcome this problem. The main() file is small because functions
defined in the same file with main() cannot be unit tested. The inline
functions are, therefore, mostly in libs. Include these libs by
including the entire C file rather than simply deleting that lib (losing
it test coverage) and moving its code into the main() file. Continue to
build the lib object file for unit tests, but eliminate the lib from the
list of dependencies in the avr-gcc build recipe.

Do not apply this strategy blindly to all libs. It is surprising which
ones require it and which do not.

For example, the lib that defines how to manipulate bits uses inline
functions. avr-gcc can still link the object file for this lib and use
the optimal assembly instructions.

I think this happens because the inline definition is shared with other
translation units *before* being converted to assembly. So it is only
libs-that-call-the-bit-manipulation-lib-on-I/O-registers that avr-gcc
cannot generate optimal assembler. It must be that this is the point in
the toolchain when the bit manipulation function is actually being
converted to assembly.