Change ReadWriteBits inline functions to macros
Inline is supposed to have the same end-result as using a macro. But
inline calls are sometimes optimized out of the `.lst`, while macros are
faithfully reproduced in the `.lst`.

avr-gcc optimizes away calls to BitIsClear(). I expect these calls to
show up like the other ReadWriteBits inline functions which show up as
expected in the `.lst`.

BitIsClear() is called in a loop `while(BitIsClear(reg, bit)), so I
expect `sbis reg, bit` followed by an `rjmp` that loops back to `sbis`.
When the bit is set, the loop ends. The compiler just eliminates the
loop entirely. Loops that translate to `sbic` are not eliminated. And
some loops with `sbis` are not optimized away. It is not clear why this
happens in some cases and not others.

I want to get to the bottom of this, but while trouble-shooting I
realized that the compiler would not optimize-away the call if I pasted
the `BitIsClear` body instead of making the call to BitIsClear. So I
changed the implementation in ReadWriteBits to macros.