Optimal assembly for SpiSlave and BiColorLED
Differentiate between hardware libs built with or without hardware defs.

If the lib has inline functions, then the translation unit that uses
those functions (the main file) needs those defs. Include the
lib-Hardware.h file in Hardware.h. List this lib with the
`inlhw_lib_src` group. This way it is still a pre-requisite for building
the avr-target. Example: BiColorLED.

If the lib does not have inline functions and it has hardware, then
its translation unit needs the hardware defs. Do not include
lib-Hardware.h in Hardware.h. List this lib with the `hw_lib_src` group.
The Makefile includes the lib-Hardware when the lib is compiled with
avr-gcc. Example: SpiSlave.

If the lib does not need any hardware defs, list it with the
`nonhw_lib_src` group. This way the -Hardware and -HardwareFake .h files
are not pre-requisites. Example: ReadWriteBits